requirements.txt
django==5.1.3
openai-whisper==20240930
pydub==0.25.1
chromadb==0.5.20
langchain-community==0.3.8